:construction: pagination

+++ src/modules/__Generic/generic.controller.ts
@@ -28,20 +28,20 @@ export class GenericController<T> {
     });
   });
 
-  // Get all items
+  // Get all items  //[🚧][🧑‍💻✅][🧪] // 🆗
   getAll = catchAsync(async (req: Request, res: Response) => {
     const result = await this.service.getAll();
 
     sendResponse(res, {
       code: StatusCodes.OK,
       data: result,
-      message: `All ${this.modelName}`,
+      message: `All ${this.modelName}s`,
       success: true,
     });
   });
 
   getAllWithPagination = catchAsync(async (req: Request, res: Response) => {
-    const filters = pick(req.query, ['_id']); // 'projectName',
+    const filters = pick(req.query, ['_id', 'title']); // 'projectName',
     const options = pick(req.query, ['sortBy', 'limit', 'page', 'populate']);
 
     const result = await this.service.getAllWithPagination(filters, options);
@@ -54,7 +54,7 @@ export class GenericController<T> {
     });
   });
 
-  // Get by ID
+  // Get by ID // [🚧][🧑‍💻✅][🧪] // 🆗
   getById = catchAsync(async (req: Request, res: Response) => {
     const id = req.params.id;
 
@@ -75,9 +75,15 @@ export class GenericController<T> {
 
   // Update by ID
   updateById = catchAsync(async (req: Request, res: Response) => {
+    if (!req.params.id) {
+      throw new ApiError(
+        StatusCodes.BAD_REQUEST,
+        `id is required for update ${this.modelName}`
+      );
+    }
     const id = req.params.id;
-    const data = req.body;
-    const updatedObject = await this.service.updateById(id, data);
+
+    const updatedObject = await this.service.updateById(id, req.body);
     if (!updatedObject) {
       throw new ApiError(
         StatusCodes.NOT_FOUND,
@@ -94,6 +100,13 @@ export class GenericController<T> {
 
   // Delete by ID
   deleteById = catchAsync(async (req: Request, res: Response) => {
+    if (!req.params.id) {
+      throw new ApiError(
+        StatusCodes.BAD_REQUEST,
+        `id is required for delete ${this.modelName}`
+      );
+    }
+
     const id = req.params.id;
     const deletedObject = await this.service.deleteById(id);
     if (!deletedObject) {
